refactor: 레일 드래그 로직을 useRailDrag 훅으로 분리
